(refactor): Move Solaris and OSX flag declaration into case statement.
Solaris requires the -m64 flag everywhere, so it must be added to the
declaration of the *_COMMON flag variables. The OSX_ARCH variable
is always preceded by the _COMMON flag variable and followed by the
OSX_FLAGS variable during module building. Therefore, add them to the
_COMMOND flag declaration within the case statement and remove all
pairs of OSX_ARCH followed by OSX_FLAGS. The Jpeg builds are left untouched.